docs: optimize for AI

+++ meshlink/src/commonMain/kotlin/ch/trancee/meshlink/model/LinkMetric.kt
@@ -7,7 +7,7 @@ package ch.trancee.meshlink.model
  */
 public data class LinkMetric(
     /** RSSI normalized to 0-255 scale (0 = unusable, 255 = excellent). */
-    public val rssiNormalized: UInt,
+    public val rssiNormalized: UByte,
     /** Whether the link supports L2CAP Connection-oriented Channels. */
     public val supportsL2CAP: Boolean,
     /** Whether the connection interval is ≤ 15 ms (short). */
@@ -17,7 +17,8 @@ public data class LinkMetric(
 ) {
     /** Composite metric value: low byte = RSSI, high bits = flags. */
     public val composite: UInt =
-        ((supportsL2CAP.bit(8) or lowLatency.bit(9) or highPower.bit(10)) shl 8) or rssiNormalized
+        ((supportsL2CAP.bit(8) or lowLatency.bit(9) or highPower.bit(10)) shl 8) or
+            rssiNormalized.toUInt()
 }
 
 private fun Boolean.bit(position: Int): UInt = if (this) 1u shl position else 0u
